Move filter box to separate file for readability
