+++ src/containers/TalkingPoints/TalkingPoints.js
@@ -2,94 +2,16 @@ import React, { Component } from 'react';
 import { Redirect } from 'react-router-dom';
 import moment from 'moment';
 import { connect } from 'react-redux';
-import { Button, Checkbox, Col, Form, Icon, Input, List, message, Modal, DatePicker, Row, Select, Skeleton, Spin, Typography } from 'antd';
+import { Button, Icon, Input, List, message, Modal, Skeleton, Spin, Typography } from 'antd';
 import get from "lodash/get"
 
 import axios from '../../_util/axios-api';
 import { authHeader } from '../../_util/auth/auth-header';
 import { slug as districtSlug } from "../../_util/district";
 import AddEditTalkingPointModal from './AddEditTalkingPointModal'
+import TalkingPointsFilterForm from './TalkingPointsFilterForm'
 
 import './TalkingPoints.module.css';
-
-class TalkingPointsFilter extends Component {
-
-    render() {
-        const filters = this.props.filters;
-        const { getFieldDecorator } = this.props.form;
-        return (
-            <Form>
-                <Row>
-                    <Col sm={24} md={12} >
-                        <Form.Item label="Creation Date">
-                            {getFieldDecorator("creationDate", {initialValue: filters && filters.creationDate})(
-                                <DatePicker.RangePicker />
-                            )}
-                        </Form.Item>
-                    </Col>
-                    <Col sm={24} md={12} >
-                        <Form.Item label="Last Updated Date">
-                            {getFieldDecorator("updatedDate", {initialValue: filters && filters.updatedDate})(
-                                <DatePicker.RangePicker />
-                            )}
-                        </Form.Item>
-                    </Col>
-                </Row>
-                <Row gutter={[{ xs: 8, sm: 16, md: 24, lg: 32 }, 20]}>
-                    <Col sm={24} md={6} >
-                        <Form.Item label="Relevance">
-                            {getFieldDecorator("scope", {initialValue: filters && filters.scope})(
-                                <Checkbox.Group>
-                                    <Checkbox value={"national"}>National</Checkbox>
-                                    <Checkbox value={"state"}>My State</Checkbox>
-                                    <Checkbox value={"district"}>My District</Checkbox>
-                            </Checkbox.Group>
-                            )}
-                        </Form.Item>
-                    </Col>
-                    <Col sm={24} md={6} >
-                        <Form.Item label="Theme">
-                        {getFieldDecorator("theme", {initialValue: filters && filters.theme})(
-                            <Select
-                                showSearch
-                                allowClear
-                                placeholder="Select a theme"
-                                optionFilterProp="children"
-                                filterOption={(input, option) => option.props.children.toLowerCase().indexOf(input.toLowerCase()) >= 0}>
-                                {this.props.themes.map((el)=>{
-                                    return (<Select.Option key={el.name} value={el.name}>{el.name}</Select.Option>)
-                                })}
-                            </Select>,
-                            )}
-                        </Form.Item>
-                    </Col>
-                    <Col sm={24} md={6} >
-                        <Form.Item label="Currently selected">
-                            {getFieldDecorator("script", {
-                                valuePropName: 'checked',
-                                initialValue: filters && filters.script
-                            })(
-                                <Checkbox value={"script"}>In Script</Checkbox>
-                            )}
-                        </Form.Item>
-                    </Col>
-                    <Col sm={24} md={6} >
-                        <Form.Item label="Author">
-                            {getFieldDecorator("author", {initialValue: filters && filters.author})(
-                                <Input placeholder="Username or email" />
-                            )}
-                        </Form.Item>
-                    </Col>
-                </Row>
-            </Form>
-        );
-    }
-}
-
-const TalkingPointsFilterForm = Form.create({
-    name: 'talking_points_filter_sort',
-    onValuesChange: (props, _, values) => props.onValuesChange(values)
-})(TalkingPointsFilter);
 
 class TalkingPoints extends Component {
